backend/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-multiselectfield==0.1.12
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
